feat: add processRequest method
