+++ packages/auth/src/guard/company.guard.ts
@@ -1,18 +1,11 @@
-import { ExecutionContext, ForbiddenException, Injectable } from '@nestjs/common'
+import { ExecutionContext, Injectable } from '@nestjs/common'
 import { AuthGuard } from '@nestjs/passport'
 import { Request } from 'express'
-import { authorizationHeader, getRequestFromContext } from '../util'
+import { processRequest } from '../util'
 
 @Injectable()
 export class CompanyGuard extends AuthGuard('company') {
   getRequest(context: ExecutionContext): Request {
-    const request = getRequestFromContext(context)
-    const token = authorizationHeader(request.headers)
-
-    if (!token) throw new ForbiddenException()
-
-    request.context = context
-
-    return request
+    return processRequest(context)
   }
 }
